Update bash_prompt to generate prompt from more variables (as opposed to the single function)
Summary:
By default, bash will count all the characters in the
$PS1 variable to determine line length·(including hidden \
characters) In order to mitigate that, we need to 'hide'
the hidden characters in \[ \] braces.  I couldn't find a good
way to do that in functions since \[ and \] kept showing up in
my prompt. Instead I changed the model of how the prompt was created
into many functions.